third_party/blink/web_tests/fast/css/dynamic-pseudo-class-expected.txt

This tests elements affected by CSS pseudo class selectors are updated after a dynamic DOM change.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(document.getElementById("orig"))["color"] is "rgb(255, 0, 0)"
PASS getComputedStyle(document.getElementById("orig"))["color"] is "rgb(0, 0, 0)"
PASS successfullyParsed is true

TEST COMPLETE

Generated content.
Original.
